🎨 Rename `generated` package to `gensql`
Rename `generated` package to `gensql` to improve clarity
on the code base. The name `generated` is too generic and
does not reveal that the generated code pertains to generated
sql code from the `sqlc` tool.
