+++ backend/scripts/generate-sqlc.dsh
@@ -1,3 +1,3 @@
 #!/usr/bin/env dagger
 
-. | generate-sqlc | export ./generated
+. | generate-sqlc | export ./gensql
